Review changes.
Mainly moved the SigmaFock self-energy to the sse_coulomb_screening module, and added the assemble_boundary_blocks to a utils module.
