--- src/quatrex/coulomb_screening/utils/__init__.py
@@ -1,7 +1,0 @@
-# Copyright 2023-2024 ETH Zurich and the QuaTrEx authors. All rights reserved.
-
-from quatrex.coulomb_screening.utils.assemble_boundary_blocks import (
-    assemble_boundary_blocks,
-)
-
-__all__ = ["assemble_boundary_blocks"]
