+++ src/quatrex/electron/__init__.py
@@ -1,8 +1,7 @@
 # Copyright 2023-2024 ETH Zurich and the QuaTrEx authors. All rights reserved.
 
 from quatrex.electron.solver import ElectronSolver
-from quatrex.electron.sse_coulomb_screening import SigmaCoulombScreening
-from quatrex.electron.sse_fock import SigmaFock
+from quatrex.electron.sse_coulomb_screening import SigmaCoulombScreening, SigmaFock
 from quatrex.electron.sse_phonon import SigmaPhonon
 from quatrex.electron.sse_photon import SigmaPhoton
